requirements.txt
fastapi>=0.109.0
uvicorn[standard]>=0.27.0
pydantic>=2.5.0
google-genai>=1.0.0
python-dotenv>=1.0.0
pytest>=8.0.0
mcp>=1.2.0
supabase>=2.0.0
yt-dlp>=2024.1.0
